Integrate analyzer into system
Fixes #33
