+++ scraper/requirements.txt
@@ -8,7 +8,7 @@ cffi==1.15.1
 charset-normalizer==3.0.1
 click==8.1.3
 confection==0.0.4
-curl-cffi==0.5.1
+curl-cffi==0.5.7
 cymem==2.0.7
 dateparser==1.1.8
 decorator==5.1.1
